Signup Form Validated + Styled
The signup modal form has been styled completely, as well as validated. When a user skips over an input, it will say to the user that the field is required. The only thing that needs to be validated is if the user clicks "signup" and the API can't sign them up due to a repeated email, to show that error.