Club Events Carousel
Added events to favorited club events carousel. Styled the event cards. Also added a template that will show when the user is not logged in on the club page, which will show featured clubs which is an array of all the clubs that is shuffled each time the page loads.