Fixed dialog issues
Two issues with the dialogs in the specific club page. One was if you clicked outside the dialog to close it, it would show two events when you click on another event to edit. Fixed that by adding the onhide function. The next one was the table data disappearing when you closed an edit event dialog. Got rid of this.createEventForm.reset();