Navbar is Now Responsive
Navbar will be responsive depending on the device type. If the device type is a tablet or a mobile phone, the navbar will collapse into a dropdown menu on the left, SNHU logo in the middle, and involveU logo on the right.

IMPORTANT. When changing screen size, you need to refresh. Reason being is because you are not able to change screen size as a normal user.