Moved Club Events to P-table
The reason I moved events to the p-table is for a couple reason. One, there was too many events to display on one page, so this is my alternative to a carousel. Two, events sometimes didn't load, so I wanted to utilize primeng's "loading" option, and so far it works flawlessly.