Show More Button for Announcements Functional
On the specific club page, the show more button in the announcements section is functional. It will open up a dialog that will show all of the announcements related to the certain club. There is no text length restriction (so no ellipsis), and it will show all of the available announcements.